+++ jcm/physics/radiation/speedy_longwave.py
