+++ jcm/physics/radiation/speedy_shortwave.py
