+++ jcm/physics/speedy/forcing.py
@@ -3,7 +3,7 @@ from jcm.physics.speedy.params import Parameters
 from jcm.physics.speedy.physics_data import ablco2_ref, PhysicsData
 from jcm.forcing import ForcingData
 from jcm.physics_interface import PhysicsState, PhysicsTendency
-from jcm.physics.speedy.shortwave_radiation import get_zonal_average_fields
+from jcm.physics.radiation.speedy_shortwave import get_zonal_average_fields
 import jax.numpy as jnp
 # linear trend of co2 absorptivity (del_co2: rate of change per year)
 del_co2   = 0.005
